explicitly check inode type before file projection
Rather than rely on O_RDWR to help us detect when the
inode we are projecting is a directory, not a file,
perform an explicit fstat(2) check instead.  This allows us
to acquire the lock using O_RDONLY, so we do not fail if
the file has a read-only permission mode.  It also allows us
to detect other inode types, such as sockets, and ignore them.